id_card_extractor_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openpyxl==3.1.2
pandas==2.2.2
python-dotenv==1.0.1
bcrypt==4.1.2
pytz==2024.1
xlsxwriter==3.2.0
google-api-python-client==2.128.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.7.1
Pillow==10.3.0
requests==2.32.3